minor fixes

+++ internal/models/user.go
@@ -6,7 +6,8 @@ import (
 )
 
 type User struct {
-	ID        int64
+	ID int
+
 	UserUID   uuid.UUID
 	Username  *string
 	AvatarUID *uuid.UUID
